third-party-service/requirements.txt
# deepface==0.0.79
# loguru==0.7.2
pika==1.3.2
python-dotenv==1.0.0
requests